move test utils to tests/utils folder
